outsourced and refactored taxonomy
